transaction_enrichment/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-extensions==3.2.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
